python typing system

+++ python/demos/2024-12/steam_utils/steam_api_usage.py
@@ -9,48 +9,50 @@ steamBD search
     是否有parent app id
 
 """
+
 import configparser
 import datetime
 import json
 
 import pandas
 import requests
-
 import save_license_to_excel
 
-api_key = '---'
-steam_id = '---'
+api_key = "---"
+steam_id = "---"
 
 
 def load_app_secure_key():
     global api_key, steam_id
     config = configparser.ConfigParser()
-    config.read('personal.ini', encoding='utf8')
+    config.read("personal.ini", encoding="utf8")
 
-    if 'steam' not in config.sections():
-        raise SystemExit('no such config')
+    if "steam" not in config.sections():
+        raise SystemExit("no such config")
 
-    steam_item = config['steam']
-    api_key = steam_item.get('api_key', '')
-    steam_id = steam_item.get('steam_id', '')
+    steam_item = config["steam"]
+    api_key = steam_item.get("api_key", "")
+    steam_id = steam_item.get("steam_id", "")
 
 
 load_app_secure_key()
 
 
 def get_supported_apilist():
-    base_url = 'https://api.steampowered.com/ISteamWebAPIUtil/GetSupportedAPIList/v1/'
-    params = {'key': api_key, }
+    base_url = "https://api.steampowered.com/ISteamWebAPIUtil/GetSupportedAPIList/v1/"
+    params = {
+        "key": api_key,
+    }
     try:
-        resp = requests.request('GET', base_url, params=params)
+        resp = requests.request("GET", base_url, params=params)
         if resp.status_code == 200:
             data = json.loads(resp.content)
 
         else:
-            raise ValueError(f'API 返回的响应码不是 200')
+            raise ValueError(f"API 返回的响应码不是 200")
 
     except Exception as e:
-        print('Fatal Error')
+        print("Fatal Error")
         print(e)
 
 
@@ -67,20 +69,19 @@ def get_owned_games(include_appinfo=True, include_played_free_games=True):
 
     """
 
-    base_url = 'https://api.steampowered.com/IPlayerService/GetOwnedGames/v1/'
-    method = 'GET'
+    base_url = "https://api.steampowered.com/IPlayerService/GetOwnedGames/v1/"
+    method = "GET"
 
-    params = {'key': api_key,
-              'steamid': steam_id}
+    params = {"key": api_key, "steamid": steam_id}
 
     if include_appinfo:
-        params['include_appinfo'] = True
+        params["include_appinfo"] = True
     if include_played_free_games:
-        params['include_played_free_games'] = True
+        params["include_played_free_games"] = True
 
-    params['include_free_sub'] = True
-    params['skip_unvetted_apps'] = False
-    params['include_extended_appinfo'] = True
+    params["include_free_sub"] = True
+    params["skip_unvetted_apps"] = False
+    params["include_extended_appinfo"] = True
 
     results = []
 
@@ -88,15 +89,15 @@ def get_owned_games(include_appinfo=True, include_played_free_games=True):
         resp = requests.request(method, base_url, params=params)
         if resp.status_code == 200:
             data = json.loads(resp.content)
-            data = data['response']
-            game_count = data['game_count']
-            game_items = data['games']  # list of dict
-            print(f'一共获取了{game_count}条数据')
+            data = data["response"]
+            game_count = data["game_count"]
+            game_items = data["games"]  # list of dict
+            print(f"一共获取了{game_count}条数据")
         else:
-            raise ValueError(f'API 返回的响应码不是 200')
+            raise ValueError(f"API 返回的响应码不是 200")
 
     except Exception as e:
-        print('Fatal Error')
+        print("Fatal Error")
         print(e)
         return results
 
@@ -104,17 +105,17 @@ def get_owned_games(include_appinfo=True, include_played_free_games=True):
     for game_item in game_items:
         # pprint.pprint(game_item)
         _item = dict()
-        _item['name'] = game_item['name']
-        _item['appid'] = game_item['appid']
-        _item['time_total'] = game_item.get('playtime_forever', 0)
-        _item['time_2weeks'] = game_item.get('playtime_2weeks', 0)
-        _item['time_total'] = round(_item['time_total'] / 60, 2)
-        _item['time_2weeks'] = round(_item['time_2weeks'] / 60, 2)
+        _item["name"] = game_item["name"]
+        _item["appid"] = game_item["appid"]
+        _item["time_total"] = game_item.get("playtime_forever", 0)
+        _item["time_2weeks"] = game_item.get("playtime_2weeks", 0)
+        _item["time_total"] = round(_item["time_total"] / 60, 2)
+        _item["time_2weeks"] = round(_item["time_2weeks"] / 60, 2)
 
-        _item['last_played'] = game_item.get('rtime_last_played', None)
-        if _item['last_played'] is not None:
+        _item["last_played"] = game_item.get("rtime_last_played", None)
+        if _item["last_played"] is not None:
             # local time
-            _item['last_played'] = datetime.datetime.fromtimestamp(_item['last_played'])
+            _item["last_played"] = datetime.datetime.fromtimestamp(_item["last_played"])
             # _item['last_played'] = _item['last_played'].strftime("%Y-%m-%d %H:%M:%S")
 
         # pprint.pprint(_item)
@@ -124,15 +125,15 @@ def get_owned_games(include_appinfo=True, include_played_free_games=True):
 
 
 def save_owned_games(res_list: list):
-    filename = 'steam_playtime_{}.xlsx'.format(save_license_to_excel.get_current_datetime())
-    df = pandas.DataFrame(res_list, columns=['appid', 'name', 'time_total', 'time_2weeks', 'last_played'])
-    df['appid'] = df['appid'].astype('int')
-    df['time_total'] = df['time_total'].astype('float')
-    df['time_2weeks'] = df['time_2weeks'].astype('float')
-    df['last_played'] = pandas.to_datetime(df['last_played'])
+    filename = "steam_playtime_{}.xlsx".format(save_license_to_excel.get_current_datetime())
+    df = pandas.DataFrame(res_list, columns=["appid", "name", "time_total", "time_2weeks", "last_played"])
+    df["appid"] = df["appid"].astype("int")
+    df["time_total"] = df["time_total"].astype("float")
+    df["time_2weeks"] = df["time_2weeks"].astype("float")
+    df["last_played"] = pandas.to_datetime(df["last_played"])
 
     df.to_excel(filename, index=False)
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     load_app_secure_key()
